web/sneaky/requirements.txt
dnspython==2.3.0
Flask==2.1.0

